test(dashboard): add api error path tests for client and update vitest config
